requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via -r requirements.in
google-api-core==2.12.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.23.2
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.3
    # via google-cloud-storage
google-cloud-storage==2.11.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.6.0
    # via google-cloud-storage
googleapis-common-protos==1.60.0
    # via google-api-core
idna==3.4
    # via requests
numpy==1.25.2
    # via -r requirements.in
pillow==10.0.0
    # via -r requirements.in
protobuf==4.24.4
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-json-logger==2.0.7
    # via -r requirements.in
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
tqdm==4.66.1
    # via -r requirements.in
urllib3==2.0.6
    # via requests
